# Improve Adler32 performance