@Xor-el 2020-05-30
I get a huge performance boost from this, from the original 946.03 MB/s to 2101.30 MB/s on my machine.
thanks for this.
merged.